+++ test/optimizer_test.cpp
@@ -33,57 +33,70 @@ RosLockGuard g_rclcpp;
 
 TEST_CASE("Optimizer doesn't fail")
 {
+  SECTION("evalControl with different critics") {
   // Settings
   int batch_size = 400;
   int time_steps = 15;
   unsigned int path_points = 50u;
   int iteration_count = 1;
   double lookahead_distance = 10.0;
+  std::vector<std::string> critics;
 
   bool consider_footprint = GENERATE(true, false);
   std::string motion_model = GENERATE("DiffDrive", "Omni");
 
-  std::vector<std::string> critics = GENERATE(
-    std::vector<std::string>{},
-    std::vector<std::string>{{"GoalCritic"}, {"GoalAngleCritic"}, {"ObstaclesCritic"},
-      {"ReferenceTrajectoryCritic"}, {"LocalGoalCritic"}},
-    std::vector<std::string>{{"GoalCritic"}},
-    std::vector<std::string>{{"GoalAngleCritic"}},
-    std::vector<std::string>{{"ObstaclesCritic"}},
-    std::vector<std::string>{{"LocalGoalCritic"}},
-    std::vector<std::string>{{"PreferForwardCritic"}},
-    std::vector<std::string>{{"ReferenceTrajectoryCritic"}},
-    std::vector<std::string>{{"PathAngleCritic"}},
-    std::vector<std::string>{{"TwirlingCritic"}}
-  );
-
   TestCostmapSettings cost_map_settings{};
+  auto costmap_ros = getDummyCostmapRos(cost_map_settings);
+  costmap_ros->setRobotFootprint(getDummySquareFootprint(0.01));
+  auto costmap = costmap_ros->getCostmap();
+
   TestPose start_pose = cost_map_settings.getCenterPose();
+  double path_step = cost_map_settings.resolution;
+
+  TestPathSettings path_settings{start_pose, path_points, path_step, path_step};
   TestOptimizerSettings optimizer_settings{batch_size, time_steps, iteration_count,
     lookahead_distance, motion_model, consider_footprint};
 
-  double path_step = cost_map_settings.resolution;
-  TestPathSettings path_settings{start_pose, path_points, path_step, path_step};
+  SECTION("Obstacles free critics test/benchmark") {
+    std::cout << "\nObstacles free critics test/benchmark" << "\n";
+    critics = GENERATE(
+      std::vector<std::string>{},
+      std::vector<std::string>{{"GoalCritic"}},
+      std::vector<std::string>{{"GoalAngleCritic"}},
+      std::vector<std::string>{{"LocalGoalCritic"}},
+      std::vector<std::string>{{"PreferForwardCritic"}},
+      std::vector<std::string>{{"ReferenceTrajectoryCritic"}},
+      std::vector<std::string>{{"PathAngleCritic"}},
+      std::vector<std::string>{{"TwirlingCritic"}}
+    );
+  }
 
-  print_info(optimizer_settings, path_settings, critics);
-  auto costmap_ros = getDummyCostmapRos(cost_map_settings);
+  SECTION("Obstacles dependent critics test/benchmark") {
+    {
+      std::cout << "\nObstacles dependent critics test/benchmark" << "\n";
+      critics = GENERATE(
+        std::vector<std::string>{{"GoalCritic"}, {"GoalAngleCritic"}, {"ObstaclesCritic"},
+          {"ReferenceTrajectoryCritic"}, {"LocalGoalCritic"}},
+        std::vector<std::string>{{"ObstaclesCritic"}}
+      );
+
+      unsigned int offset = 4;
+      unsigned int obstacle_size = offset * 2;
+
+      unsigned char obstacle_cost = 250;
+
+      auto [obst_x, obst_y] = cost_map_settings.getCenterIJ();
+
+      obst_x = obst_x - offset;
+      obst_y = obst_y - offset;
+      addObstacle(costmap, {obst_x, obst_y, obstacle_size, obstacle_cost});
+    }
+  }
+
+  printInfo(optimizer_settings, path_settings, critics);
   auto node = getDummyNode(optimizer_settings, critics);
   auto parameters_handler = std::make_unique<mppi::ParametersHandler>(node);
   auto optimizer = getDummyOptimizer(node, costmap_ros, parameters_handler.get());
-
-  // setup costmap
-  auto costmap = costmap_ros->getCostmap();
-  {
-    costmap_ros->setRobotFootprint(getDummySquareFootprint(0.01));
-    const unsigned int obst_center_x = cost_map_settings.cells_x / 2;
-    const unsigned int obst_center_y = cost_map_settings.cells_y / 2 + 1;
-    TestObstaclesSettings obs_settings_1{obst_center_x - 4, obst_center_y, 4, 255};
-    TestObstaclesSettings obs_settings_2{obst_center_x + 4, obst_center_y, 4, 255};
-    TestObstaclesSettings obs_settings_3{obst_center_x + 12, obst_center_y, 4, 255};
-    addObstacle(costmap, obs_settings_1);
-    addObstacle(costmap, obs_settings_2);
-    addObstacle(costmap, obs_settings_3);
-  }
 
   // evalControl args
   auto pose = getDummyPointStamped(node, start_pose);
@@ -98,6 +111,8 @@ TEST_CASE("Optimizer doesn't fail")
 #else
   REQUIRE_NOTHROW(optimizer.evalControl(pose, velocity, path, dummy_goal_checker));
 #endif
+
+  }
 }
